// src/api/ps.cust.api/res/ps.cust.psychologist.detail.res.ts
import { PsCustMainRes } from './ps.cust.main.res';

export class PsCustPsychologistDetailRes extends PsCustMainRes {
  custCode: string;
  custType: string;
  custTypeName: string;
  firstname: string;
  lastname: string;
  registerCode: string;
  status: string;
  mobileNo: string;
  experience: string;
  about: string;
}