// src/api/ps.cust.api/data/ps.cust.service.req.data.ts
export class PsCustServiceReqData {
  recId: number;
  custCode: string;
  custType: string;
  spCustCode: string;
  spCustType: string;
  serviceType: string;
  serviceDate: Date;
  startDatetime: Date;
  endDatetime: Date;
  duration: number;
  status: string;
  price: number;
  totalPrice: number;
  createdDatetime: Date;
}